requirements.txt
python-dotenv==1.0.0
ccxt==4.3.0
requests==2.31.0
websocket-client==1.6.0
numpy==1.24.3
pandas==2.0.3
colorama==0.4.6
rich==13.6.0
psutil==5.9.6
